+++ packages/design-system/src/molecules/sidebar/index.tsx
@@ -13,19 +13,36 @@ export const Sidebar: React.FC<SidebarProps & PropsWithChildren> = ({
 }) => {
   const isMobile = useBreakpointValue({ base: true, xl: false })
   const theme = useThemeV2()
-  const isNative = Platform.OS !== 'web'
-  if (isMobile) {
-    return (
-      <View>
-        <MobileSidebar {...props} />
-        {isNative ? <ScrollView>{children}</ScrollView> : children}
-      </View>
-    )
-  }
   return (
-    <View style={[container.row, { backgroundColor: theme.background.grey }]}>
-      <DesktopSidebar {...props} />
-      <View style={[container.flex1]}>{children}</View>
-    </View>
+    <>
+      {Platform.select({
+        native: (
+          <View>
+            <MobileSidebar {...props} />
+            <ScrollView>{children}</ScrollView>
+          </View>
+        ),
+        web: (
+          <>
+            {isMobile ? (
+              <View>
+                <MobileSidebar {...props} />
+                {children}
+              </View>
+            ) : (
+              <View
+                style={[
+                  container.row,
+                  { backgroundColor: theme.background.grey },
+                ]}
+              >
+                <DesktopSidebar {...props} />
+                <View style={[container.flex1]}>{children}</View>
+              </View>
+            )}
+          </>
+        ),
+      })}
+    </>
   )
 }
