+++ packages/design-system/src/molecules/sidebar/types.ts
@@ -4,6 +4,11 @@ export interface SidebarFooterOptionProps {
   title: string
   link: string
 }
+
+export interface SidebarSection {
+  id: string
+  options: SidebarOptionProps[]
+}
 export interface LogoSidebar {
   desktop: React.ReactNode
   mobileHeader: React.ReactNode
@@ -13,7 +18,8 @@ export interface SidebarProps {
   logo: LogoSidebar
   header?: React.ReactNode
   selectedOptionId?: string
-  options: SidebarOptionProps[][]
+  // options: SidebarOptionProps[][]
+  sections: SidebarSection[]
   withConnectWallet?: boolean
   footerOptionsOnMobile?: SidebarFooterOptionProps[]
   footerOnDesktop?: React.ReactNode
